# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is an attendance tracking system API using RFID",
        "title": "IoT Attendance System API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
        "/dashboard/history": {
            "get": {
                "description": "Retrieves complete attendance history for all students",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "dashboard"
                ],
                "summary": "Get attendance history",
                "responses": {
                    "200": {
                        "description": "Attendance history retrieved successfully",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/main.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/main.AttendanceHistory"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "500": {
                        "description": "Failed to fetch attendance history",
                        "schema": {
                            "$ref": "#/definitions/main.Response"
                        }
                    }
                }
            }
        },
        "/dashboard/present": {
            "get": {
                "description": "Lists all students currently present in the facility",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "dashboard"
                ],
                "summary": "Get currently present students",
                "responses": {
                    "200": {
                        "description": "Current students retrieved successfully",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/main.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/main.StudentPresence"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "500": {
                        "description": "Failed to fetch current students",
                        "schema": {
                            "$ref": "#/definitions/main.Response"
                        }
                    }
                }
            }
        },
        "/student/{rfid}/attendance": {
            "get": {
                "description": "Retrieves attendance history for a specific student",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "students"
                ],
                "summary": "Get student attendance history",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Student RFID number",
                        "name": "rfid",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Attendance records retrieved successfully",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/main.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/main.AttendanceRecord"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "404": {
                        "description": "Student not found",
                        "schema": {
                            "$ref": "#/definitions/main.Response"
                        }
                    },
                    "500": {
                        "description": "Failed to fetch attendance records",
                        "schema": {
                            "$ref": "#/definitions/main.Response"
                        }
                    }
                }
            }
        },
        "/walkin/{rfid}": {
            "post": {
                "description": "Records when a student enters using their RFID card",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "attendance"
                ],
                "summary": "Register student walk-in",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Student RFID number",
                        "name": "rfid",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Walk-in registered successfully",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "Student not found",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Failed to register walk-in",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/walkout/{rfid}": {
            "post": {
                "description": "Records when a student exits using their RFID card",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "attendance"
                ],
                "summary": "Register student walk-out",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Student RFID number",
                        "name": "rfid",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Walk-out registered successfully",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "No active walk-in found for walk-out",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "Student not found",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Failed to register walk-out",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "main.AttendanceHistory": {
            "type": "object",
            "properties": {
                "duration": {
                    "type": "string"
                },
                "in_time": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "out_time": {
                    "type": "string"
                },
                "rfid": {
                    "type": "string"
                }
            }
        },
        "main.AttendanceRecord": {
            "type": "object",
            "properties": {
                "duration": {
                    "type": "string"
                },
                "in_time": {
                    "type": "string"
                },
                "out_time": {
                    "type": "string"
                },
                "rfid": {
                    "type": "string"
                }
            }
        },
        "main.Response": {
            "type": "object",
            "properties": {
                "data": {},
                "message": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                }
            }
        },
        "main.StudentPresence": {
            "type": "object",
            "properties": {
                "department": {
                    "type": "string"
                },
                "in_time": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "rfid": {
                    "type": "string"
                }
            }
        }
    }
}